+++ cogs/test.py
@@ -1,0 +1,12 @@
+import discord
+from discord.ext import commands
+
+class test(commands.Cog):
+
+    def __init__(self, bot):
+        self.bot = bot
+
+
+def setup(bot):
+    bot.add_cog(test(bot))
+print("hi")
